Vista jugadores en poanel admin

+++ src/main/java/edu/martin/dida/proyectofinciclo/admin/ControladorAdmin.java
@@ -4,17 +4,40 @@
  */
 package edu.martin.dida.proyectofinciclo.admin;
 
+import POJO.Player;
+import POJO.Team;
 import POJO.Usuario;
+import Utilidades.SendMessage;
+import com.cloudinary.Cloudinary;
+import com.cloudinary.utils.ObjectUtils;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import edu.martin.dida.proyecto.conexion.Conexion;
+import edu.martin.dida.proyecto.conexion.PlayersDAO;
 import edu.martin.dida.proyecto.conexion.UsuariosDAO;
+import edu.martin.dida.proyectofinciclo.ControladorLogin.ControladorLogin;
 import edu.martin.dida.proyectofinciclo.ControladorLogin.ControladorRegistro;
+import edu.martin.dida.proyectofinciclo.inicio.ControladorInicio;
+import java.io.BufferedWriter;
+import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Base64;
 import java.util.Base64.Encoder;
 import java.util.List;
+import java.util.Map;
 import java.util.ResourceBundle;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.regex.Matcher;
@@ -25,10 +48,13 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
+import javafx.scene.control.Slider;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.text.Text;
+import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import javax.swing.JFrame;
 import javax.swing.JOptionPane;
@@ -40,8 +66,13 @@ import javax.swing.JOptionPane;
 public class ControladorAdmin {
     
     
+    
+    
     @FXML
     private Button btnMinimize;
+    
+    // <editor-fold defaultstate="collapsed" desc="User">
+
     
     @FXML
     private TableView tableuser;
@@ -65,18 +96,70 @@ public class ControladorAdmin {
     private Text lblCountUser;
 
     @FXML
-    private TableView tableadmin;
-    
-    
-    
-    
+    private TableView tableadmin;    
     boolean decryptBool = true;
+    
+                // </editor-fold>
+    
+    // <editor-fold defaultstate="collapsed" desc="Player">
+
+    @FXML
+    TextField txtJersey;  
+    
+    @FXML
+    TextField txtName; 
+    
+    @FXML
+    TextField txtPosition; 
+    
+    @FXML
+    TextField txtWeight; 
+    
+    @FXML
+    TextField txtHeight;  
+    
+    @FXML
+    Slider slider;
+    
+    @FXML
+    TextField txtAge; 
+    
+    @FXML
+    TextField txtDraft; 
+    
+    @FXML
+    TextField txtCollege;   
+    
+    @FXML
+    TextField txtNationality;    
+    
+    @FXML
+    TableView TablePlayerUser;
+    
+    @FXML
+    TextField txtImage;
+        
+    @FXML
+    TextField txtPoints;
+        
+    @FXML
+    TextField txtAssist;
+        
+    @FXML
+    TextField txtRebbounds;
+        
+    @FXML
+    public ComboBox comboTeam;
+                // </editor-fold>
+
     int id;
-    
+    int idplayer;
+    PlayersDAO playerDao;
+    ObservableList<String> items = FXCollections.observableArrayList();
 
-
     
-    public void close(){
+    public void close() throws IOException{
+        UsuariosDAO.updateStatusOffline(ControladorLogin.nombre);
         System.exit(0);
     }
     
@@ -86,7 +169,8 @@ public class ControladorAdmin {
     }
     
     
-    
+    // <editor-fold defaultstate="collapsed" desc="UserTab">
+
     
     public void cargarUser() throws IOException {
         ObservableList<Usuario> users = FXCollections.observableArrayList();
@@ -102,39 +186,49 @@ public class ControladorAdmin {
         tableadmin.setItems(users);
     }
     
-
-    
-    
     public void editUser() throws IOException{
         Usuario user = (Usuario) tableuser.getSelectionModel().getSelectedItem();
         Usuario admin = (Usuario) tableadmin.getSelectionModel().getSelectedItem();
+            
 
         if(!(user == null)){
-            txtUser.setText(user.getUser());
-            txtEmail.setText(user.getEmail());
-            txtPassword.setText(encrypt(user.getPassword()));
-            
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
-            String date = user.getDateBirth();
-            LocalDate localDate = LocalDate.parse(date, formatter);
-            txtDate.setValue(localDate);
-            
-            if(user.getPermisos().equals("user")){
-                checkAdmin.setSelected(false);
+            if(!(UsuariosDAO.getStatus(user.getUser()) == true)){
+                txtUser.setText(user.getUser());
+                txtEmail.setText(user.getEmail());
+                txtPassword.setText(encrypt(user.getPassword()));
+
+                DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+                String date = user.getDateBirth();
+                LocalDate localDate = LocalDate.parse(date, formatter);
+                txtDate.setValue(localDate);
+
+                if(user.getPermisos().equals("user")){
+                    checkAdmin.setSelected(false);
+                }
+                id = user.getId();
+            }else{
+            JOptionPane.showMessageDialog(new JFrame(), "This user is online", "Error", JOptionPane.ERROR_MESSAGE);        
             }
-            id = user.getId();
+
         
          }else if(user==null && admin!=null){
-            txtUser.setText(admin.getUser());
-            txtEmail.setText(admin.getEmail());
-            txtPassword.setText(encrypt(admin.getPassword()));
+                if(!(UsuariosDAO.getStatus(admin.getUser()) == true)){
+                    txtUser.setText(admin.getUser());
+                txtEmail.setText(admin.getEmail());
+                txtPassword.setText(encrypt(admin.getPassword()));
+
+
+
+                if(admin.getPermisos().equals("admin")){
+                    checkAdmin.setSelected(true);
+                }
+                id = admin.getId();
+                }else{
+                    JOptionPane.showMessageDialog(new JFrame(), "This user is online", "Error", JOptionPane.ERROR_MESSAGE);        
+
+                }
+
             
-            
-            
-            if(admin.getPermisos().equals("admin")){
-                checkAdmin.setSelected(true);
-            }
-            id = admin.getId();
             
             
             
@@ -167,8 +261,8 @@ public class ControladorAdmin {
                 }
 
 
-                ControladorRegistro.userDAO.saveEdit(user);
-
+                ControladorRegistro.userDAO.saveEditUser(user);
+                numberUser();
                 cargarAdmin();
                 cargarUser();
                 clearFieldsUser();
@@ -189,11 +283,26 @@ public class ControladorAdmin {
         Usuario admin = (Usuario) tableadmin.getSelectionModel().getSelectedItem();
 
         if(!(user == null)){
-            ControladorRegistro.userDAO.deleteUser(user);
-            cargarUser();
+            if(!(UsuariosDAO.getStatus(user.getUser()) == true)){
+                ControladorRegistro.userDAO.deleteUser(user);
+                cargarUser();
+                cargarPlayers();
+                numberUser();
+            }else{
+            JOptionPane.showMessageDialog(new JFrame(), "This user is online", "Error", JOptionPane.ERROR_MESSAGE);        
+            }
         }else if(user==null && admin!=null){
-            ControladorRegistro.userDAO.deleteUser(user);
-            cargarAdmin();
+            if(!(UsuariosDAO.getStatus(admin.getUser()) == true)){
+                ControladorRegistro.userDAO.deleteUser(admin);
+                cargarAdmin();
+                cargarPlayers();
+                numberUser();
+            }else{
+                
+                JOptionPane.showMessageDialog(new JFrame(), "This user is online", "Error", JOptionPane.ERROR_MESSAGE);        
+            }
+
+            
         }else if(admin == null && user == null){
             JOptionPane.showMessageDialog(new JFrame(), "You must select a user", "Error", JOptionPane.ERROR_MESSAGE);        
         }
@@ -203,6 +312,174 @@ public class ControladorAdmin {
     public void numberUser() throws IOException{
         lblCountUser.setText(String.valueOf(ControladorRegistro.userDAO.countUsers()));
     }
+    
+    public void getJson() throws IOException{
+        Path path = Paths.get(System.getProperty("user.dir") + "/TheBull.json");
+
+        Gson gson = new GsonBuilder().setPrettyPrinting().create();
+        BufferedWriter bw = Files.newBufferedWriter(path);
+        
+        String email = ControladorRegistro.userDAO.emailAdmin(ControladorLogin.nombre);
+        List<Usuario> listuser = ControladorRegistro.userDAO.usersJson();
+        
+        for(int i = 0; i < listuser.size(); i++){
+            bw.write(gson.toJson(listuser.get(i)) + "\n");
+        }
+        
+        bw.flush();
+        bw.close();
+        
+        File file = new File(System.getProperty("user.dir") + "/TheBull.json");
+        
+        
+        ExecutorService service = Executors.newFixedThreadPool(4);
+            service.submit(new Runnable() {
+                public void run() {
+                        SendMessage.sendFile(email, file);
+                            }
+                        });
+            JOptionPane.showMessageDialog(new JFrame(), "Email with users sent to " + email, "Info", JOptionPane.INFORMATION_MESSAGE);        
+    }
+    
+                // </editor-fold>
+
+    public void cargarPlayers() throws IOException {
+        playerDao = new PlayersDAO();
+
+        ObservableList players = FXCollections.observableArrayList();
+        List<Player> player = playerDao.buscarPlayer();
+        players.addAll(player);
+        TablePlayerUser.setItems(players);
+    }
+    
+    
+    public ObservableList<String> chargeTeamsCombo() throws IOException{
+        try(Connection conexion = Conexion.getConnection()){
+            String sql = "SELECT DISTINCT name FROM teams";
+            PreparedStatement prepstmt = conexion.prepareStatement(sql);
+            items.removeAll(items);
+            ResultSet rs = prepstmt.executeQuery();
+            while(rs.next()){
+                items.add(rs.getString("name"));
+            }
+            return items;
+        }catch(SQLException e){
+            Utilidades.Logger.logInfo(e.getMessage() , 2);
+            return null;
+
+        }
+    }
+    
+    public void editPlayer(){
+        Player player = (Player) TablePlayerUser.getSelectionModel().getSelectedItem();
+        
+        if(player == null){
+            JOptionPane.showMessageDialog(new JFrame(), "You must select a player", "Error", JOptionPane.ERROR_MESSAGE);        
+        }else{
+            txtJersey.setText(String.valueOf(player.getPlayerJersey()));
+            txtName.setText(player.getPlayerName());
+            txtPosition.setText(player.getPlayerPosition());
+            txtWeight.setText(player.getPlayerWeight());
+            txtHeight.setText(player.getPlayerHeight());
+            txtAge.setText(String.valueOf(player.getPlayerAge()));
+            txtDraft.setText(player.getPlayerDraft());
+            txtCollege.setText(player.getPlayerCollege());
+            txtNationality.setText(player.getPlayerNationality());
+            txtPoints.setText(String.valueOf(player.getPlayerPoints()));
+            txtRebbounds.setText(String.valueOf(player.getPlayerRebbounds()));
+            txtAssist.setText(String.valueOf(player.getPlayerAssist()));
+            txtImage.setText(player.getPlayerImage());        
+            comboTeam.setValue(player.getPlayerTeam());
+            idplayer = player.getPlayerId();
+        }
+    }
+    
+    
+    public void insertPlayer() throws IOException{
+         Player player = new Player();
+        playerDao = new PlayersDAO();
+        String user = null;
+        if (txtJersey.getText().isEmpty() || txtName.getText().isEmpty() || txtPosition.getText().isEmpty() || txtWeight.getText().isEmpty() || txtHeight.getText().isEmpty()
+                || txtAge.getText().isEmpty() || txtDraft.getText().isEmpty() || txtCollege.getText().isEmpty() || txtNationality.getText().isEmpty() || txtPoints.getText().isEmpty()
+                || txtAssist.getText().isEmpty() || txtRebbounds.getText().isEmpty() || comboTeam.getValue()==null || txtImage.getText().isEmpty()){
+            JOptionPane.showMessageDialog(new JFrame(), "All fields must be covered", "Error", JOptionPane.ERROR_MESSAGE);        
+        }else{
+                if(txtImage.getText().indexOf("cloudinary") != -1 || txtImage.getText().indexOf("cdn.nba") != -1){
+                    player.setPlayerJersey(Integer.parseInt(txtJersey.getText()));
+                    player.setPlayerName(txtName.getText());
+                    player.setPlayerPosition(txtPosition.getText());
+                    player.setPlayerWeight(txtWeight.getText() +"kg");
+                    player.setPlayerHeight(txtHeight.getText() + "cm");
+                    player.setPlayerTeam((String) comboTeam.getSelectionModel().getSelectedItem());
+                    player.setPlayerAge(Integer.parseInt(txtAge.getText()));
+                    player.setPlayerDraft(txtDraft.getText());
+                    player.setPlayerCollege(txtCollege.getText());
+                    player.setPlayerNationality(txtNationality.getText());
+                    player.setPlayerPoints(Double.parseDouble(txtPoints.getText()));
+                    player.setPlayerRebbounds(Double.parseDouble(txtRebbounds.getText()));
+                    player.setPlayerAssist(Double.parseDouble(txtAssist.getText()));
+                    player.setPlayerImage(txtImage.getText());  
+                    player.setPlayerId(idplayer);
+                    
+                    ControladorRegistro.userDAO.saveEditPlayer(player, user);
+                        cargarPlayers();
+                        idplayer = 0;
+                        JOptionPane.showMessageDialog(new JFrame(), "Inserted", "Error", JOptionPane.ERROR_MESSAGE);        
+                }else{
+                    try {
+                        player.setPlayerJersey(Integer.parseInt(txtJersey.getText()));
+                        player.setPlayerName(txtName.getText());
+                        player.setPlayerPosition(txtPosition.getText());
+                        player.setPlayerWeight(txtWeight.getText() +"kg");
+                        player.setPlayerHeight(txtHeight.getText() + "cm");
+                        player.setPlayerTeam((String) comboTeam.getSelectionModel().getSelectedItem());
+                        player.setPlayerAge(Integer.parseInt(txtAge.getText()));
+                        player.setPlayerDraft(txtDraft.getText());
+                        player.setPlayerCollege(txtCollege.getText());
+                        player.setPlayerNationality(txtNationality.getText());
+                        player.setPlayerPoints(Double.parseDouble(txtPoints.getText()));
+                        player.setPlayerRebbounds(Double.parseDouble(txtRebbounds.getText()));
+                        player.setPlayerAssist(Double.parseDouble(txtAssist.getText()));
+                        player.setPlayerImage(dinary(txtImage.getText()));
+                        player.setPlayerId(idplayer);
+                        user = ControladorRegistro.userDAO.searchUserPlayer(player.getPlayerTeam());
+                        
+                        ControladorRegistro.userDAO.saveEditPlayer(player, user);
+                        cargarPlayers();
+                        idplayer = 0;
+                        JOptionPane.showMessageDialog(new JFrame(), "Inserted", "Info", JOptionPane.ERROR_MESSAGE);        
+                    } catch (SQLException ex) {
+                        Logger.getLogger(ControladorAdmin.class.getName()).log(Level.SEVERE, null, ex);
+                    }
+                    
+                }
+        }
+
+            }
+    
+    public void deletePlayer() throws IOException{
+        Player player = (Player) TablePlayerUser.getSelectionModel().getSelectedItem();
+        
+        if(player == null){
+            JOptionPane.showMessageDialog(new JFrame(), "You must select a player", "Error", JOptionPane.ERROR_MESSAGE);        
+   
+        }else{
+            ControladorRegistro.userDAO.deletePlayer(player);
+            JOptionPane.showMessageDialog(new JFrame(), "Player deleted", "Error", JOptionPane.INFORMATION_MESSAGE);       
+            cargarPlayers();
+
+        }
+    }
+    
+        public void getImage(){
+            FileChooser f = new FileChooser();
+            f.getExtensionFilters().add(new FileChooser.ExtensionFilter("PNG FILES", "*.png"));
+            File file = f.showOpenDialog(null);
+            if(!(file == null)){
+                txtImage.setText(file.getAbsolutePath());
+            }
+    }
+    
     
     
     public void desencriptar(){
@@ -248,5 +525,19 @@ public class ControladorAdmin {
         txtEmail.clear();
         txtDate.setValue(null);
         txtPassword.clear();
+    }
+    
+        private String dinary(String url) throws IOException{
+        Cloudinary cloudinary = new Cloudinary(ObjectUtils.asMap(
+        "cloud_name", "daolhlyb6",
+        "api_key", "853471969531956",
+        "api_secret", "yQIl1DHYJbh0Gm3td56uD7d66ts",
+        "secure", true));
+        
+        File file = new File(url);
+        
+        Map uploadresult = cloudinary.uploader().upload(file, ObjectUtils.emptyMap());
+        String a = (String) uploadresult.get("url");
+        return a;
     }
 }
